% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R
\name{backtest.indep}
\alias{backtest.indep}
\title{Backtest Intra-day VaR (Independence Hypothesis)}
\usage{
backtest.indep(vio, K)
}
\arguments{
\item{vio}{A (grid_point) x (number of observations) matrix drawn from the violation process curves.}

\item{K}{The maximal lagged autocorrelation considered for the independence test.}
}
\value{
The p-value of the test statistics.
}
\description{
backtest.indep function backtests the independence hypothesis for the intra-day VaR curves forecasts.
}
\details{
Given the violation process \eqn{Z_i^\tau(t)} at the quantile \eqn{\tau} obtained by using \code{\link{var.vio}}, for \eqn{t \in[0,1]} and \eqn{1\leq i\leq N}, the function provides an approximate p-value for a test of the null hypothesis of \eqn{H_0}: \eqn{Z_i^\tau(t)} being IID along \eqn{i}.\cr
To test the null hypothesis, the function applies the portmanteau test statistic proposed by Kokoszka, Rice, Shang, (2017),\cr
\eqn{V_{N,K}=N\sum_{h=1}^K||\hat{\gamma}_{h,Z}||^2}, \cr
where \eqn{K} is a pre-set maximum lag length, and the autocovariance function \eqn{\hat{\gamma}_{h,Z}(t,s)=\frac{1}{N}\sum_{i=1}^{N-h}[Z_i(t)-\bar{Z}_i(t)][Z_{i+h}(s)-\bar{Z}(s)]}, for \eqn{||\cdot ||} is the \eqn{L^2} norm, and \eqn{\bar{Z}(t)=1/N\sum_{i=1}^N Z_i(t)}.
}
\examples{
# generate discrete evaluations of the FGARCH(1,1) process.
grid_point = 50; N = 200
yd = dgp.fgarch(grid_point, N, "garch")
yd = yd$garch_mat
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd, M=2)
basis_est = ba$basis
# fit the curve data and the conditional volatility by using an FGARCH(1,1) model with M=1.
fd = fda::Data2fd(argvals=seq(0,1,len=50),y=yd,fda::create.bspline.basis(nbasis=32))
y_inp = basis.score(fd, basis_est[,1])
garch11_est = est.fGarch(y_inp)
diag_garch = diagnostic.fGarch(garch11_est, basis_est[,1], yd)
# get the in-sample fitting of conditional variance.
sigma_fit = diag_garch$sigma2[,1:N]
error_fit = diag_garch$eps
# get in-sample intra-day VaR curve by assuming a point-wisely Gaussian distributed error term.
var_obj = var.forecast(yd, sigma_fit, error_fit, quantile_v=0.01, Method="normal")
intra_var = var_obj$intraday_VaR
# obtain the violation curves.
intra_vio = var.vio(yd, intra_var)

# backtesting the Independence Hypothesis for the violation curve.
backtest.indep(vio=intra_vio, K=5)

}
\references{
Christoffersen, P. (2010). Backtesting. Encyclopedia of Quantitative Finance, Wiley. \cr
Kokoszka, P., Rice, G., Shang, H. L. (2017). Inference for the autocovariance of a functional time series under conditional heteroscedasticity. Journal of Multivariate Analysis, 162, 32-50. \cr
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via intra-Day return curves. International Journal of Forecasting.
}
\seealso{
\code{\link{var.forecast}} \code{\link{var.vio}} \code{\link{backtest.unbias}}
}
