% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R
\name{backtest.unbias}
\alias{backtest.unbias}
\title{Backtest Intra-day VaR (Unbiasedness Hypothesis)}
\usage{
backtest.unbias(vio, tau)
}
\arguments{
\item{vio}{A (grid_point) x (number of observations) matrix drawn from the violation process curves.}

\item{tau}{The nominal/true quantile of the VaR curves.}
}
\value{
The p-value of the test statistics.
}
\description{
backtest.unbias function backtests the unbiasedness hypothesis for the intra-day VaR curve forecasts.
}
\details{
Given the violation process \eqn{Z_i^\tau(t)} at the quantile \eqn{\tau} obtained by using \code{\link{var.vio}}, the function computes an approximate p-value of a test of the null hypothesis \eqn{H_0}: \eqn{E(Z_i^\tau(t)-\tau)=0}, for all \eqn{t \in[0,1]}, \eqn{1\leq i\leq N}. The test statistics,\cr
\eqn{T_N=N|| \bar{Z}(t)-\tau||^2} is employed, where \eqn{||\cdot ||} is the \eqn{L^2} norm, and \eqn{\bar{Z}(t)=1/N\sum_{i=1}^N Z_i(t)}.
}
\examples{
# generate discrete evaluations of the FGARCH(1,1) process.
grid_point = 50; N = 200
yd = dgp.fgarch(grid_point, N, "garch")
yd = yd$garch_mat
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd, M=2)
basis_est = ba$basis
# fit the curve data and the conditional volatility by using a FGARCH(1,1) model with M=1.
fd = fda::Data2fd(argvals=seq(0,1,len=grid_point),y=yd,fda::create.bspline.basis(nbasis=32))
y_inp = basis.score(fd, basis_est[,1])
garch11_est = est.fGarch(y_inp)
diag_garch = diagnostic.fGarch(garch11_est, basis_est[,1], yd)
# get the in-sample fitting of conditional variance.
sigma_fit = diag_garch$sigma2[,1:N]
error_fit = diag_garch$eps
# get in-sample intra-day VaR curve by assuming a point-wisely Gaussian distributed error term.
var_obj = var.forecast(yd, sigma_fit, error_fit, quantile_v=0.01, Method="normal")
intra_var = var_obj$intraday_VaR
# compute the violation curves.
intra_vio = var.vio(yd, intra_var)

# backtesting the Unbiasedness Hypothesis for the violation curve.
backtest.unbias(vio=intra_vio, tau=0.01)

}
\references{
Christoffersen, P. (2010). Backtesting. Encyclopedia of Quantitative Finance, Wiley. \cr
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via intra-Day return curves. International Journal of Forecasting.
}
\seealso{
\code{\link{var.forecast}} \code{\link{var.vio}} \code{\link{backtest.indep}}
}
