% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_return.R
\name{basis.est}
\alias{basis.est}
\title{Estimating non-negative basis functions}
\usage{
basis.est(yd, M, type)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N discrete evaluation curves.}

\item{M}{The number/order of basis functions, by setting \eqn{M <= 10}.}

\item{type}{A string to switch estimation method: "poly" - exponential and Bernstein functions; "tfpca" - truncated functional principal components; "fsnn" - sparse and non-negative functional principal components; "tpf" - truncated predictive factors.}
}
\value{
List of objects:

basis: a (grid_point) x (M) matrix containing discrete evaluations of M data-driven bases; when "poly" is applied, the basis contains a list with 'exp' for exponential and 'bern' for Bernstein.

tve: the total variations explained by data-driven bases, not applied for polynomials.
}
\description{
basis.est function generates/estimates non-negative basis functions, so that they can be used to estimate a Functional GARCH model subsequently.
}
\examples{
# generate discrete evaluations of the FGARCH process.
yd = dgp.fgarch(50, 100, "garch")
yd = yd$garch_mat

# decompose the first truncated non-negative functional principal component.
dt = basis.est(yd, M=1, "tfpca")
tbasis = dt$basis
tve = dt$tve

}
\references{
Cerovecki, C., Francq, C., Hormann, S., Zakoian, J. M. (2019). Functional GARCH models: The quasi-likelihood approach and its applications. Journal of Econometrics. 209(2), 353-375. <doi:10.1016/j.jeconom.2019.01.006>.\cr
Rice, G., Wirjanto, T., Zhao, Y. (2021) Exploring volatility of crude oil intra-day return curves: a functional GARCH-X model. MPRA Paper No. 109231. <https://mpra.ub.uni-muenchen.de/109231>.\cr
}
