% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_return.R
\name{basis.fsnn}
\alias{basis.fsnn}
\title{Sparse and Non-Negative Functional Principal Component Analysis}
\usage{
basis.fsnn(yd, M)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N curves.}

\item{M}{The number/order of basis functions, by default \eqn{M = 10}.}
}
\value{
List of objects:

basis: a (grid_point) x (M) matrix containing discrete evaluations of M sparse and non-negative functional principal components.

tve: the total variations explained by each sparse and non-negative functional principal component.
}
\description{
basis.fsnn function extracts sparse and non-negative functional principal components from the squared process of intra-day returns.
}
\examples{
# generate discrete evaluations of the FARCH process.
yd = dgp.farch(50,100)
yd_arch = yd$arch_mat

# decompose the first sparse and non-negative functional principal component.
basis.fsnn(yd_arch,M=1)

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Functional GARCH-X model with an application to forecasting crude oil return curves. Working Paper.
}
\seealso{
\code{\link{basis.pp}} \code{\link{basis.tfpca}} \code{\link{basis.tpf}}
}
