% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_return.R
\name{basis.pp}
\alias{basis.pp}
\title{Positive Polynominals}
\usage{
basis.pp(grid_point, M)
}
\arguments{
\item{grid_point}{The number of grid point in each curve observation.}

\item{M}{The number/order of basis functions.}
}
\value{
List of objects:

exp: A (grid_point) x (M) matrix drawn from M Exponential basis functions.

bern: A (grid_point) x (M) matrix drawn from M Bernstein basis functions.
}
\description{
basis.pp function returns positive vectors drawn from exponential and Bernstein functions.
}
\examples{
#generate discrete evaluations of exponential and Bernstein basis functions with order one.
ppb = basis.pp(50, 1)
exp = ppb$exp
bern = ppb$bern
}
\seealso{
\code{\link{basis.tfpca}} \code{\link{basis.fsnn}} \code{\link{basis.tpf}}
}
