% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{basis.score}
\alias{basis.score}
\title{Projection Scores}
\usage{
basis.score(fdata, basis)
}
\arguments{
\item{fdata}{The functional data object with N paths.}

\item{basis}{The M-dimensional basis functions.}
}
\value{
A (N) x (M) matrix containing functional scores.
}
\description{
basis.score function returns functional scores by projecting the squared process onto some given basis functions.
}
\examples{
# generate discrete evaluations of iid curve data.
edata = dgp.fiid(50,100)
# smooth discrete data into functional curves.
fd = fda::Data2fd(argvals=seq(0,1,len = 50),y=edata,fda::create.bspline.basis(nbasis = 32))
bern = basis.pp(50,2)$bern

# get functional scores by projecting the squared process onto Bernstein basis.
y_inp = basis.score(fd,bern)
}
\seealso{
\code{\link{est.fArch}} \code{\link{est.fGarch}} \code{\link{est.fGarchx}} \code{\link{diagnostic.fGarch}}
}
