% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R, R/intraday_return.R
\name{basis.tfpca}
\alias{basis.tfpca}
\title{Truncated Functional Pricipal Component Analysis}
\usage{
basis.tfpca(yd, M)

basis.tfpca(yd, M)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N discrete evaluation curves.}

\item{M}{The number/order of basis functions, by default \eqn{M = 10}.}
}
\value{
List of objects:

basis: a (grid_point) x (M) matrix containing discrete evaluations of M truncated functional principal components.

tve: the total variations explained by each truncated functional principal component.

List of objects:

basis: a (grid_point) x (M) matrix containing discrete evaluations of M truncated functional principal components.

tve: the total variations explained by each truncated functional principal component.
}
\description{
basis.tfpca function extracts truncated non-negative functional principal components from the squared process of conditionally heteroscedastic curves, presumably, so that they may be modelled with a Functional GARCH model subsequently.

basis.tfpca function extracts truncated non-negative functional principal components from the squared process of conditionally heteroscedastic curves, presumably, so that they may be modelled with a Functional GARCH model subsequently.
}
\examples{
# generate discrete evaluations of the FARCH process.
fd = dgp.farch(50,100)
fd_arch = fd$arch_mat

# decompose the first truncated non-negative functional principal component.
dt = basis.tfpca(fd_arch,M=1)
tbasis = dt$basis
tve = dt$tve

# generate discrete evaluations of the FARCH process.
fd = dgp.farch(50,100)
fd_arch = fd$arch_mat

# decompose the first truncated non-negative functional principal component.
dt = basis.tfpca(fd_arch,M=1)
tbasis = dt$basis
tve = dt$tve

}
\references{
Cerovecki, C., Francq, C., Hormann, S., Zakoian, J. M. (2019). Functional GARCH models: The quasi-likelihood approach and its applications. Journal of econometrics, 209(2), 353-375.

Cerovecki, C., Francq, C., Hormann, S., Zakoian, J. M. (2019). Functional GARCH models: The quasi-likelihood approach and its applications. Journal of econometrics, 209(2), 353-375.
}
\seealso{
\code{\link{basis.pp}} \code{\link{basis.fsnn}} \code{\link{basis.pf}}

\code{\link{basis.pp}} \code{\link{basis.fsnn}} \code{\link{basis.tpf}}
}
