% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_return.R
\name{basis.tpf}
\alias{basis.tpf}
\title{Truncated Predictive Factors}
\usage{
basis.tpf(yd, M, a, h)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N discrete evaluation curves.}

\item{M}{The number/order of basis functions, by default \eqn{M = 10}.}

\item{a}{The regularization parameter with recommended values of 0.73 - 0.75.}

\item{h}{The forecasting horizon.}
}
\value{
List of objects:

basis: a (grid_point) x (M) matrix containing discrete evaluations of M non-negative predictive factors.

tve: the total variations explained by each non-negative predictive factor.
}
\description{
basis.tpf function extracts non-negative predictive factors from the squared process of conditionally heteroscedastic curves, presumably, so that they may be modelled with a Functional GARCH model subsequently.
}
\examples{
# generate discrete evaluations of the FARCH process.
yd = dgp.farch(50,100)
yd_arch = yd$arch_mat

# decompose the first two non-negative predictive factors.
basis.tpf(yd_arch,M=2,a=0.75,h=1)

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Functional GARCH-X model with an application to forecasting crude oil return curves. Working Paper.
}
\seealso{
\code{\link{basis.pp}} \code{\link{basis.tfpca}} \code{\link{basis.fsnn}}
}
