% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp.farch}
\alias{dgp.farch}
\title{Data Generating Process - Functional ARCH Process}
\usage{
dgp.farch(grid_point, N, beta_par = NULL)
}
\arguments{
\item{grid_point}{The number of grid point in each curve observation.}

\item{N}{The sample size.}

\item{beta_par}{The kernel coefficient function in the conditional volatility equation. If it is missing, the default function "\eqn{16 * t * (1-t) * s * (1-s)}" is used, for \eqn{t\in[0,1]} and \eqn{s\in[0,1]}.}
}
\value{
List of generated processes:

arch_mat: FARCH sequences, where the finite realization of curves are stored in columns,

sigma_mat: Conditional volatility sequences,  where the finite realization of curves are stored in columns.
}
\description{
dgp.farch function generates functional curve data following the functional ARCH(1) process.
}
\details{
\eqn{x_i(t)} is generated to follow an FARCH(1) process, such that,\cr
\eqn{x_i(t)=\sigma_i(t)\varepsilon_i(t)}, \eqn{t \in [0,1]},\cr
\eqn{\sigma_i^2(t)=\omega(t)+\int \alpha(t,s) x^2_{i-1}(s)ds},\cr
where the innovation \eqn{\varepsilon_i(t)} follows an Ornstein–Uhlenbeck process \code{\link{dgp.fiid}}, and the kernel coefficients \eqn{\omega(t)=0.1t(1-t)}, \eqn{\alpha(t,s)=12t(1-t)s(1-s)}.
}
\examples{
# generate discrete evaluations of 100 FARCH curves that each curve is realized on 50 grid points.
yd = dgp.farch(50,100)
yd_arch = yd$arch_mat

# smooth discrete data into functional curves.
fd = fda::Data2fd(argvals=seq(0,1,len = 50),y=yd_arch,fda::create.bspline.basis(nbasis = 32))
}
\references{
Hormann, S., Horvath, L., Reeder, R. (2013). A functional version of the ARCH model. Econometric Theory, 29(2), 267-288.
}
\seealso{
\code{\link{dgp.fiid}} \code{\link{dgp.fgarch}}
}
