% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{diagnostic.fGarch}
\alias{diagnostic.fGarch}
\title{Diagnostic information derived from the estimation}
\usage{
diagnostic.fGarch(params, basis, yd, p = 0, q = 1, xd = NULL)
}
\arguments{
\item{params}{List of model paramters: d, As, Bs (and Gs).}

\item{basis}{The M-dimensional basis functions for the projection.}

\item{yd}{A grid_point x N matrix drawn from N functional curves.}

\item{p}{The order of the depedence on past squared observations. If it is missing, p=0.}

\item{q}{The order of the depedence on past volatilities. If it is missing, q=1.}

\item{xd}{A grid_point x N matrix drawn from N covariate X curves. The default value is NULL.}
}
\value{
List of parameters:

eps: a grid_point x N matrix containing fitted residuals.

sigma2: a grid_point x (N+1) matrix that the first N columns are the fitted conditional variance, the N+1 column is the predicted conditional variance.

yfit: a grid_point x N matrix drawn from N fitted intra-day price curves.

kernel_op: estimated kernel coefficient operators.
}
\description{
diagnostic.fGarch function provides the estimation parameters that can be used as the inputs for a diagnostic purpose.
}
\examples{
# generate discrete evaluations of the FARCH process.
yd = dgp.fgarch(grid_point=50, N=200, "garch")
yd = yd$garch_mat
fdy = fda::Data2fd(argvals=seq(0,1,len=50),y=yd,fda::create.bspline.basis(nbasis=32))
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd, M=2)
basis_est = ba$basis
# fit the curve data with an FARCH(1) model.
y_inp = basis.score(fdy, basis_est)
arch_est = est.fArch(y_inp)

# get parameters for diagnostic checking.
diag_arch  = diagnostic.fGarch(arch_est, basis_est, yd)

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Tests for conditional heteroscedasticity of functional data. Journal of Time Series Analysis.
}
\seealso{
\code{\link{basis.score}} \code{\link{est.fArch}} \code{\link{est.fGarch}} \code{\link{est.fGarchx}} \code{\link{gof.fgarch}}
}
