% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete_gof.R
\name{fun_hetero}
\alias{fun_hetero}
\title{Test for Conditional Heteroscedasticity of Functional Data}
\usage{
fun_hetero(yd, K = NULL, stat_Method, pplot = NULL)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N intra-day return curves.}

\item{K}{The lag autocorrelation coefficients served for the conditional heteroscedasticity test. If it is missing, a default value "K=20" is used.}

\item{stat_Method}{A string to indicate which test will be implemented: "norm" - \eqn{V_{N,K}}; "functional" - \eqn{M_{N,K}}. Or if it is missing, the "functional"-type method is used.}

\item{pplot}{an optional argument to compute and plot the P-values (either "norm" or "functional") as a function of K, K=1,2,...,20. If pplot=1, the p-values will be computed and figured; if it is missing, this step will be skipped.}
}
\value{
List of objects:

stats: the test statistic.

p_value: the P_value.
}
\description{
fun_hetero applies a test of the null hypothesis that the objective functional curve data is not conditionally heteroscedastic, small p-values suggest that the curves exhibit conditional heteroscedasticity.
}
\details{
Given the objective curve data \eqn{x_i(t)}, for \eqn{1\leq i \leq N}, \eqn{t\in[0,1]}, the test aims at distinguishing the hypotheses:\cr
\eqn{H_0}: the sequence \eqn{x_i(t)} is IID; \cr
\eqn{H_1}: the sequence \eqn{x_i(t)} is conditionally heteroscedastic. Two portmanteau type statistics are applied: \cr
the norm-based statistic: \eqn{V_{N,K}=N\sum_{h=1}^K\hat{\rho}_h^2}, where \eqn{\hat{\rho}_h} is the sample autocorrelation of the time series \eqn{||x_1||^2,\dots,||x_N||^2}, and \eqn{K} is a pre-set maximum lag length.\cr
the fully functional statistic \eqn{M_{N,K}=N\sum_{h=1}^K||\hat{\gamma}_h||^2}, where the autocovariance kernel \eqn{\hat{\gamma}_{h}(t,s)=\frac{1}{N}\sum_{i=1}^{N-h}[x_i^2(t)-\bar{x}^2(t)][x^2_{i+h}(s)-\bar{x}(s)]}, for \eqn{||\cdot ||} is the \eqn{L^2} norm, and \eqn{\bar{x}(t)=1/N\sum_{i=1}^N x^2_i(t)}.
}
\examples{
# generate discrete evaluations of the iid curves under the null hypothesis.
yd = dgp.fiid(50,100)

# test the conditional heteroscedasticity.
fun_hetero(yd, K=5,"functional")

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Tests for conditional heteroscedasticity of functional data. Journal of Time Series Analysis.
}
\seealso{
\code{\link{sample_data}}
}
