% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday_return.R
\name{intra.return}
\alias{intra.return}
\title{Intra-day Return Curves}
\usage{
intra.return(yd)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N intra-day price curves.}
}
\value{
List of return data:

idr: the IDR.

cidr: the CIDR.

ocidr: the OCIDR.
}
\description{
intra.return function calculates three types of intra-day return curves: intra-day log return (IDR), cumulative intra-day log return (CIDR), and overnight cumulative intra-day log return (OCIDR).
}
\details{
Suppose that \eqn{P_i(t)} denotes the intra-day price curves, for \eqn{1\leq i \leq N} and \eqn{t\in[0,1]}, we then have,\cr
Intra-day log returns: \eqn{x_i(t)=\log P_i(t) - \log P_i(t-\Delta)}, where \eqn{\Delta} is the frequency at intra-day grid points;
Cumulative intra-day log returns: \eqn{X_i(t)=\log P_i(t) - \log P_i(0)};
Overnight cumulative intra-day log returns: \eqn{x_i(t)=\log P_i(t) - \log P_{i-1}(1)}.
}
\examples{
# generate intra-day price curve data for the FGARCH process.
yd = dgp.fgarch(50, 100, "garch")
yd = yd$garch_mat

# calculate discrete data drawn from N intra-day return curves.
fcurve = intra.return(yd)
idr = fcurve$idr
cidr = fcurve$cidr
ocidr = fcurve$ocidr

}
\references{
Gabrys, R., Horvath, L., Kokoszka, P. (2010). Tests for error correlation in the functional linear model. Journal of the American Statistical Association, 105(491), 1113-1125.\cr
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via Intra-Day Return Curves. To appear in International Journal of Forecasting.
}
