% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for Analysis}
\format{
a 390 by 788 matrix that each column containing the raw intra-day price discretely observed at daily basis.
}
\usage{
sample_data
}
\description{
A dataset of S&P 500 intra-day price at one-minute frequence from 9:30 a.m. to 4:00 p.m.. The sample ranges from 02/January/2015 to 29/March/2018, and trading days with missing intra-day observations have been eliminated from the sample.
}
\examples{
\dontrun{
# load the data
yd = as.matrix(sample_data)
grid_point = dim(yd)[1]
N = dim(yd)[2]

# smooth data into functional objects by using B-spline bases.
fd = fda::Data2fd(y=yd, argvals=seq(0,1,len = grid_point), fda::create.bspline.basis(nbasis = 32))
}
}
\seealso{
\code{\link{intra.return}}
}
\keyword{datasets}
