% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R
\name{var.forecast}
\alias{var.forecast}
\title{Predict Daily/intra-day Value-at-Risk}
\usage{
var.forecast(yd, sigma_pred, error_fit, quantile_v, Method)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N functional curves.}

\item{sigma_pred}{The predicted conditional standard deviation curve obtained by using the functional ARCH/GARCH model.}

\item{error_fit}{The residual obtained from the functional ARCH/GARCH model.}

\item{quantile_v}{The percentile of the VaR forecasts.}

\item{Method}{A string to indicate which method will be implemented to estimate the quantile of the error: "normal" - assuming errors following a normal distribution; "bootstrap" - bootstrapping the empirical distribution from the historical data; "evt" - fitting the error with a generalized Pareto distribution.}
}
\value{
List of objects:

day_VaR: the daily VaR.

day_ES: the daily expected shortfall.

intraday_VaR: the intra-day VaR curves.
}
\description{
var.forecast function forecasts the daily VaR and intra-day VaR curves according to intra-day return curves.
}
\details{
This function uses a two-step approach to forecast intra-day Value-at-Risk, in formula that\cr
\eqn{VaR_{i+1}^\tau(t)=\hat{\sigma}_{i+1}(t)\hat{\varepsilon}^\tau(t)}, for \eqn{1\leq i \leq N}, \eqn{t\in[0,1]}, and the percentile \eqn{\tau \in [0,1]},\cr
where the forecats of conditional standard deviation \eqn{\hat{\sigma}_{i+1}(t)} can be obtained by using \code{\link{est.fArch}}, \code{\link{est.fGarch}}, or \code{\link{est.fGarchx}}. Note that when \eqn{t=1}, \eqn{VaR_{i+1}^\tau(1)} is the forecast of daily VaR.
}
\examples{
# generate discrete evaluations of the FGARCH(1,1) process.
grid_point = 50; N = 200
yd = dgp.fgarch(grid_point, N)
yd = yd$garch_mat
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd,M=2)
basis_est = ba$basis
# fit the curve data and the conditional volatility by using an FGARCH(1,1) model with M=1.
fd = fda::Data2fd(argvals=seq(0,1,len = 50),y=yd,fda::create.bspline.basis(nbasis = 32))
y_inp = basis.score(fd,as.matrix(basis_est[,1]))
garch11_est = est.fGarch(y_inp)
diag_garch = diagnostic.fGarch(garch11_est,basis_est[,1],yd)
sigma_fit = diag_garch$sigma2[,1:N]
error_fit = diag_garch$eps

# get in-sample intra-day VaR curve by assuming a point-wisely Gaussian distributed error term.
exr=var.forecast(yd,sigma_fit,error_fit,quantile_v=0.01,Method="normal")
exr$intraday_VaR

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via intra-Day return curves. International Journal of Forecasting.
}
\seealso{
\code{\link{basis.tfpca}} \code{\link{basis.score}} \code{\link{est.fGarch}} \code{\link{diagnostic.fGarch}}
}
