% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R
\name{var.vio}
\alias{var.vio}
\title{Violation Process}
\usage{
var.vio(yd, var_curve)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N functional curves.}

\item{var_curve}{A (grid_point) x (number of observations) matrix storing the forecasts of intra-day VaR curves.}
}
\value{
A violation process of the intra-day return curves based on the forecasts of intra-day VaR curves.
}
\description{
var.vio function returns a violation process for the intra-day VaR curves.
}
\details{
Given the intra-day return curves \eqn{x_i(t)}, and the forecasts of intra-day VaR curves \eqn{\widehat{VaR}_i^\tau(t)} obtained from \code{\link{var.forecast}}, the violation process \eqn{Z_i^\tau(t)} can be defined as, \cr
\eqn{Z_i^\tau(t)=I(x_i(t)<\widehat{VaR}_i^\tau(t))}, for \eqn{1\leq i \leq N}, \eqn{t\in[0,1]}, and \eqn{\tau \in [0,1]},\cr
where \eqn{I(\cdot)} is an indicator function.
}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via intra-day return curves. International Journal of Forecasting. <doi:10.1016/j.ijforecast.2019.10.006>.\cr
}
\seealso{
\code{\link{var.forecast}} \code{\link{var.backtest}}
}
