% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est.fGarch}
\alias{est.fGarch}
\title{Fit a Functional GARCH(p,q) model}
\usage{
est.fGarch(y_vec, p = 1, q = 1)
}
\arguments{
\item{y_vec}{(number of observations) x (number of basis functions) matrix of projections of squared observations.}

\item{p}{order of the depedence on past volatilities.}

\item{q}{order of the depedence on past squared observations.}
}
\value{
List of model paramters:

d: d Parameter vector, for intercept function \eqn{\delta}. See Aue et. al 2017/Cerovecki2019

As: A Matrices, for \eqn{\alpha} operators. See Aue et. al 2017/Cerovecki2019

Bs: B Matrices, for \eqn{\beta} operators. See Aue et. al 2017/Cerovecki2019
}
\description{
This function allows you to fit a funcGARCH(p,q) using QMLE, given the matrix of squared projections onto a finite number of basis functions.
}
\examples{
# generate discrete evaluations of the FARCH process.
yd = dgp.fgarch(grid_point=50, N=200, "arch")
yd = yd$garch_mat
fd = fda::Data2fd(argvals=seq(0,1,len=50),y=yd,fda::create.bspline.basis(nbasis=32))
# extract data-driven basis functions through the truncated FPCA method.
basis_est = basis.est(yd, M=2, "tfpca")$basis

# fit a FGARCH(1,1) model with the projection M=1.
y_inp = basis.score(fd, basis_est[,1])
garch11_est = est.fGarch(y_inp)

}
\references{
Aue, A., Horvath, L., and F. Pellatt, D. (2017) Functional Generalized Autoregressive Conditional Heteroskedasticity. J. Time Ser. Anal., 38: 3- 21. doi: 10.1111/jtsa.12192.\cr
Cerovecki, Clement & Francq, Christian & Hormann, Siegfried & Zakoian, Jean-Michel, 2019. "Functional GARCH models: The quasi-likelihood approach and its applications," Journal of Econometrics, Elsevier, vol. 209(2), pages 353-375.
}
\seealso{
\code{\link{basis.score}} \code{\link{est.fArch}} \code{\link{est.fGarchx}} \code{\link{diagnostic.fGarch}}
}
