% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est.fGarchx}
\alias{est.fGarchx}
\title{Estimate Functional GARCH-X Model}
\usage{
est.fGarchx(y_vec, x_vec)
}
\arguments{
\item{y_vec}{A (number of observations) x (number of basis functions) matrix of projections of squared observations.}

\item{x_vec}{A (number of observations) x (number of basis functions) matrix of projections of covariate X.}
}
\value{
List of model paramters:

d: d Parameter vector, for intercept function \eqn{\delta}.

As: A Matrices, for \eqn{\alpha} operators.

Bs: B Matrices, for \eqn{\beta} operators.

Gs: G Matrices, for \eqn{\gamma} operators.
}
\description{
est.fGarchx function estimates the Functional GARCH-X model by using the Quasi-Maximum Likelihood Estimation method.
}
\details{
This function estimates the Functional GARCH-X model:\cr
\eqn{y_i(t)=\sigma_i(t)\varepsilon_i(t)}, for \eqn{t \in [0,1]} and \eqn{1\leq i \leq N},\cr
\eqn{\sigma_i^2(t)=\omega(t)+\int \alpha(t,s) y^2_{i-1}(s)ds+\int \beta(t,s) \sigma^2_{i-1}(s)ds + \int \theta(t,s) x^2_{i-1}(s)ds},\cr
where \eqn{x_i(t)} is an exogenous variable.
}
\examples{
# generate discrete evaluations of the FARCH and FGARCH processes.
yd = dgp.farch(grid_point=50, N=200)
yd = yd$arch_mat
xd = dgp.fgarch(grid_point=50, N=200)
xd = xd$garch_mat
fdy = fda::Data2fd(argvals=seq(0,1,len = 50),y=yd,fda::create.bspline.basis(nbasis = 32))
fdx = fda::Data2fd(argvals=seq(0,1,len = 50),y=xd,fda::create.bspline.basis(nbasis = 32))
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd,M=2)
basis_est = ba$basis

# fit a FGARCH-X model with the projection M=1.
y_inp = basis.score(fdy,as.matrix(basis_est[,1]))
x_inp = basis.score(fdx,as.matrix(basis_est[,1]))
garchx_est = est.fGarchx(y_inp, x_inp)

}
\references{
Cerovecki, C., Francq, C., Hormann, S., Zakoian, J. M. (2019). Functional GARCH models: The quasi-likelihood approach and its applications. Journal of econometrics, 209(2), 353-375.\cr
Rice, G., Wirjanto, T., Zhao, Y. (2020). Functional GARCH-X Model with an Application to Forecasting Crude Oil Return Curves. Working paper.
}
\seealso{
\code{\link{basis.score}} \code{\link{est.fArch}} \code{\link{est.fGarch}} \code{\link{diagnostic.fGarch}}
}
