% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hete_gof.R
\name{gof.fgarch}
\alias{gof.fgarch}
\title{Goodness-of-fit Test for Functional ARCH/GARCH Model}
\usage{
gof.fgarch(yd, basis, model, K = NULL, pplot = NULL, max_eval = 10000)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N discrete evaluation curves.}

\item{basis}{The M-dimensional basis functions.}

\item{model}{A string to indicate which model will be estimated: "arch" - FARCH(1); "garch" - FGARCH(1,1).}

\item{K}{The statistic will be based on K lag autocorrelation coefficients. If it is missing, a default value "K=20" is used.}

\item{pplot}{An optional argument to compute and plot the P-values as a function of K, K=1,2,...,20. If pplot=1, the p-values will be computed and figured; if it is missing, this step will be skipped.}

\item{max_eval}{The maximum number of evaluations of the optimization function.}
}
\value{
The P_value of the \eqn{M_{N,K}} statistic.
}
\description{
gof.fgarch function tests the goodness-of-fit of functional ARCH/GARCH models by accounting for the effect of functional GARCH parameter estimation.
}
\details{
The test statistic used is as the same as the \eqn{M_{N,K}} statistic in \code{\link{fun_hetero}}. However, the asymptotic distribution is adjusted to account for the estimation effect, because the model residual depends on the joint asymptotics of the innovation process and the estimated parameters. We assume that the kernel parameters are consistently estimated by the Least Squares method proposed in Aue et al. (2017). Then, the asymptotic distribution of the statistic \eqn{M_{N,K}} is given in Theorem 3.1 in Rice et al. (2020).
}
\examples{
\dontrun{
# generate discrete evaluations of the FGARCH process.
grid_point=50; N=200
yd = dgp.fgarch(grid_point, N, "garch")
yd = yd$garch_mat

# extract data-driven basis functions through the truncated FPCA method.
basis_est = basis.est(yd, M=2, "tfpca")$basis

# test the adequacy of the FARCH(1) model.
gof.fgarch(yd, basis_est[,1], "arch", K=5)
}
}
\references{
Aue, A., Horvath, L., F. Pellatt, D. (2017). Functional generalized autoregressive conditional heteroskedasticity. Journal of Time Series Analysis. 38(1), 3-21. <doi:10.1111/jtsa.12192>.\cr
Rice, G., Wirjanto, T., Zhao, Y. (2020). Tests for conditional heteroscedasticity of functional data. Journal of Time Series Analysis. 41(6), 733-758. <doi:10.1111/jtsa.12532>.\cr
}
\seealso{
\code{\link{basis.est}} \code{\link{est.fGarch}} \code{\link{diagnostic.fGarch}}
}
