% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_var.R
\name{var.vio}
\alias{var.vio}
\title{Violation Process}
\usage{
var.vio(yd, var_curve)
}
\arguments{
\item{yd}{A (grid_point) x (number of observations) matrix drawn from N functional curves.}

\item{var_curve}{A (grid_point) x (number of observations) matrix storing the forecasts of intra-day VaR curves.}
}
\value{
A violation process of the intra-day return curves based on the forecasts of intra-day VaR curves.
}
\description{
var.vio function returns a violation curve for the intra-day VaR curves.
}
\details{
Given the intra-day return curves \eqn{x_i(t)}, and the forecasts of intra-day VaR curves \eqn{\widehat{VaR}_i^\tau(t)} obtained from \code{\link{var.forecast}}, the violation process \eqn{Z_i^\tau(t)} can be defined as, \cr
\eqn{Z_i^\tau(t)=I(x_i(t)<\widehat{VaR}_i^\tau(t))}, for \eqn{1\leq i \leq N}, \eqn{t\in[0,1]}, and \eqn{\tau \in [0,1]},\cr
where \eqn{I(\cdot)} is an indicator function.
}
\examples{
# generate discrete evaluations of the FGARCH(1,1) process.
grid_point = 50; N = 200
yd = dgp.fgarch(grid_point, N, "garch")
yd = yd$garch_mat
# extract data-driven basis functions through the truncated FPCA method.
ba = basis.tfpca(yd, M=2)
basis_est = ba$basis
# fit the curve data and the conditional volatility by using an FGARCH(1,1) model with M=1.
fd = fda::Data2fd(argvals=seq(0,1,len=grid_point),y=yd,fda::create.bspline.basis(nbasis=32))
y_inp = basis.score(fd,basis_est[,1])
garch11_est = est.fGarch(y_inp)
diag_garch = diagnostic.fGarch(garch11_est, basis_est[,1], yd)
# get the in-sample fitting of conditional variance.
sigma_fit = diag_garch$sigma2[,1:N]
error_fit = diag_garch$eps
# get in-sample intra-day VaR curve by assuming a point-wisely Gaussian distributed error term.
var_obj = var.forecast(yd, sigma_fit, error_fit, quantile_v=0.01, Method="normal")
intra_var = var_obj$intraday_VaR

# compute the violation curve.
var.vio(yd, intra_var)

}
\references{
Rice, G., Wirjanto, T., Zhao, Y. (2020). Forecasting Value at Risk via intra-Day return curves. International Journal of Forecasting.
}
\seealso{
\code{\link{var.forecast}}
}
